// packages/frontend/src/lib/api/queries/haravan-orders/type.ts
export enum OrderChannel {
  SPE = 'SPE',
  LZD = 'LZD',
  TTS = 'TTS',
}
export enum OrderPaymentMethod {
  COD = 'COD',
  PLATFORM = 'PLATFORM',
}

export enum OrderPaymentStatus {
  'PAID' = 'PAID',
  'UNPAID' = 'UNPAID',
}

export enum OrderStatus {
  'PROCESSING' = 'PROCESSING',
  'CANCELLED' = 'CANCELLED',
  'COMPLETED' = 'COMPLETED',
  'RETURNED' = 'RETURNED',
}

export enum OrderFinancialStatus {
  'PENDING' = 'pending',
  'AUTHORIZED' = 'authorized',
  'PARTIALLY_PAID' = 'partially_paid',
  'PAID' = 'paid',
  'PARTIALLY_REFUNDED' = 'partially_refunded',
  'REFUNDED' = 'refunded',
  'VOIDED' = 'voided',
}

export enum OrderHistoryStatus {
  'CREATED' = 'CREATED', // Đã khởi tạo
  'CLOSED' = 'CLOSED', // Đã đóng
  'CANCELLED' = 'CANCELLED', // Đã hủy
  'CONFIRMED' = 'CONFIRMED', // Đã xác nhận

  'UPDATED' = 'UPDATED', // Cập nhật thông tin

  'NOTFULFILLED' = 'NOTFULFILLED', // Chờ giao hàng
  'FULFILLED' = 'FULFILLED', // Đã giao hàng
  'NULL' = 'NULL', // Chờ giao hàng
  'PARTIAL' = 'PARTIAL', // Giao một phần
  'RESTOCKED' = 'RESTOCKED', // Đã trả hàng
}

export enum OrderFulfilmentStatus {
  WAITING_FOR_PICKUP = 'WAITING_FOR_PICKUP', // Chờ lấy hàng
  PICKING_UP = 'PICKING_UP', // Đang đi lấy
  DELIVERING = 'DELIVERING', // Đang giao hàng
  DELIVERED = 'DELIVERED', // Đã giao hàng
  DELIVERY_CANCELED = 'DELIVERY_CANCELED', // Hủy giao hàng
  RETURNED = 'RETURNED', // Chuyển hoàn
  NOT_DELIVERED_YET = 'NOT_DELIVERED_YET', // Chưa giao hàng
  CUSTOMER_NOT_FOUND = 'CUSTOMER_NOT_FOUND', // Không gặp khách
  WAITING_FOR_RETURN = 'WAITING_FOR_RETURN', // Chờ chuyển hoàn
  INCOMPLETE = 'INCOMPLETE', // Chưa hoàn thành
  DELIVERY_FAILED = 'DELIVERY_FAILED', // Lỗi giao hàng
  PICKUP_FAILED = 'PICKUP_FAILED', // Lấy hàng thất bại
}
export const ORDER_FULFILMENT_STATUS_LABEL = {
  [OrderFulfilmentStatus.WAITING_FOR_PICKUP]: 'Chờ lấy hàng',
  [OrderFulfilmentStatus.PICKING_UP]: 'Đang đi lấy',
  [OrderFulfilmentStatus.DELIVERING]: 'Đang giao hàng',
  [OrderFulfilmentStatus.DELIVERED]: 'Đã giao hàng',
  [OrderFulfilmentStatus.DELIVERY_CANCELED]: 'Hủy giao hàng',
  [OrderFulfilmentStatus.RETURNED]: 'Chuyển hoàn',
  [OrderFulfilmentStatus.NOT_DELIVERED_YET]: 'Chưa giao hàng',
  [OrderFulfilmentStatus.CUSTOMER_NOT_FOUND]: 'Không gặp khách',
  [OrderFulfilmentStatus.WAITING_FOR_RETURN]: 'Chờ chuyển hoàn',
  [OrderFulfilmentStatus.INCOMPLETE]: 'Chưa hoàn thành',
  [OrderFulfilmentStatus.DELIVERY_FAILED]: 'Lỗi giao hàng',
  [OrderFulfilmentStatus.PICKUP_FAILED]: 'Lấy hàng thất bại',
}
// ...
export interface LineItem2 {
  fulfillable_quantity: number
  fulfillment_service: any
  fulfillment_status: string
  grams: number
  id: number
  price: number
  price_original: number
  price_promotion: number
  product_id: number
  quantity: number
  requires_shipping: boolean
  sku: string
  title: string
  variant_id: number
  variant_title: string
  vendor: string
  type: string
  name: string
  gift_card: boolean
  taxable: boolean
  tax_lines: any
  product_exists: boolean
  barcode: string
  properties: Property[]
  applied_discounts: any[]
  total_discount: number
  image: Image
  not_allow_promotion: boolean
  ma_cost_amount: number
  actual_price: number
  discount_allocations: any
  serviceCode: string
  dwPrice: number
  dwTitle: string
  unitName: string
  vatRate: number
}

interface Property {
  name: string
  value: string
}

interface Image {
  src: string
}